+++ Modifier.php
@@ -24,6 +24,7 @@ use League\Uri\Contracts\PathInterface;
 use League\Uri\Contracts\UriAccess;
 use League\Uri\Contracts\UriInterface;
 use League\Uri\Exceptions\SyntaxError;
+use League\Uri\IPv4Calculators\IPv4Calculator;
 use Psr\Http\Message\UriFactoryInterface;
 use Psr\Http\Message\UriInterface as Psr7UriInterface;
 use Stringable;
@@ -218,9 +219,9 @@ final class Modifier implements Stringable, JsonSerializable, UriAccess
      *
      * @see https://url.spec.whatwg.org/#concept-ipv4-parser
      */
-    public function normalizeIPv4(?IPv4Normalizer $normalizer = null): self
+    public function normalizeIPv4(?IPv4Calculator $calculator = null): self
     {
-        $hostIp = Host::fromUri($this->uri)->toIPv4($normalizer);
+        $hostIp = Host::fromUri($this->uri)->toIPv4($calculator);
 
         return match (true) {
             null === $hostIp => $this,